Fix development color mode preference
This is a sed gone wrong.
It was unintentionally changed in 15ca95a31b51682f868e486e9f07c64d785c3f4b
which broke functionality.

Change it to a SwitchPreferenceCompat instead of changing it back
to a SwitchPreference, as all the other preferences in the
development section are one as well.

Change-Id: Id276d3b67c103b83ba01a69ff2991246c69efc69
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>