Settings: Validate that WFC carrier app is present if defined
Carriers such as AT&T, Bell and Verizon require WfcActivation for WFC
to function. It is possible however for the app to not be present on
the system, but still be defined in CarrierConfig.

This will lead to the preference being visible, but crashes when the
user attempts to enable WFC.

Change-Id: Id8ef943097a712655b7597b75b25bb0c115986c6